Working early version Utility AI

+++ ZeusTest/Assets/Romain/Considerations/ResourceConsideration.cs
@@ -8,9 +8,9 @@ public class ResourceConsideration : Consideration
     [SerializeField] private AnimationCurve responseCurve;
     public override float ScoreConsideration(NPCController _NPCController)
     {
-        score = responseCurve.Evaluate(Mathf.Clamp01(_NPCController.stats.resource /
+        score = responseCurve.Evaluate(Mathf.Clamp01((float)_NPCController.stats.resource /
                                                      _NPCController.Inventory.MaxCapacity));
-        Debug.Log(Mathf.Clamp01(_NPCController.stats.resource /
+        Debug.Log(Mathf.Clamp01((float)_NPCController.stats.resource /
                                 _NPCController.Inventory.MaxCapacity));
         return score;
     }
